Remove remnants of "inner" insalan project
Back when we did not know if we wanted to have one repository for
everything or not, the root of the django project was created inside
of `/insalan` on the repository. We later decided to move everything up,
but left the remnants of the inner project, leading to confusion as to
where what happens. This commit solves the aforementionned confusion by
destroying all remnants of the inner project.

Signed-off-by: Amélie "Lymkwi" González <lymkwi@vulpinecitrus.info>