📚 Begin uniformization and translation
Begin translating the entirety of the website to French as default, with
an English translation provided alongside.

The apps translated so far are:
 - `tournament`
 - `langate`

The translation is currently provided in `locale` at the root of the
repository. In order to avoid pushing compiled localization files, a new
entry is added to gitignore.

Signed-off-by: Amélie "Lymkwi" González <lymkwi@vulpinecitrus.info>