💶 Introduce the product category hook system
This commit introduces the product category hook system, which allows
applications to create a class derived from an interface that implements
callbacks for payment success and failure handling.

A module can simply inherit the `PaymentHooks` class in one of its
classes, and register it with the `PaymentCallbackSystem` class.

An example is provided with the tournament class, which now implements a
way to validate a pending registration with a payment.