📅 Add product and tournament expiry/origin times
Add fields to a Product and a Tournament that reflect the starting
moment at which they can be bought/you can register, and the time limit
to do so.

Also add foreign keys to a tournament to store the products it creates,
so as to not created them multiple times. These products are updated
whenever their Tournament gets changed, but not the other way back (that
will be done later).

The API also enforces time constraints, as well as whether or not a
tournament is announced.