🏆 Add a method to compute the validity of a Team
Add a method that computes whether or not a team can be considered valid
at any given point, based on quotas of people who paid.

For now, the parameter `n` used is the number of people who joined the
team. Very soon, it will be changed by the number of slots in a team
(which is tournament dependent).