Implement proper use of prepare_hook + redirect
Properly redirect the user… by not redirecting them. Redirects in Django
only really work within the same domain, so instead we return the URL in
a JSON reply, and will let the frontend do its magic.

At the same time, we move the call to the prepare_hook after creation of
the transaction object (for validation purposes), but before the intent
creation on HelloAsso's side. If the hooks return `False`, then
something is wrong, and we cannot proceed with payment, needing instead
to delete the Transaction.